Add initial project skeleton for floyd version 0.0.1.
This sets up the repo as a python project with all the skeleton
logic I use these days, including a venv with the appropriate
set of developer tools.

There is nothing whatsoever specific about this project (i.e.,
no project-specific code or logic) except for the project name
and description.