Land initial v0.1.0 working version.
This copies over working code from glop v0.7.0, with as few changes
as possible to be able to run things (though the code was reformatted
w/ ruff, and copyright headers were updated).

This also adds a couple of sample grammars that can be used for
hand-testing things for now, but does not add the bulk of the
tests, as I want to revisit how I do those.

`check` and `lint` are unhappy, and the `coverage` numbers are
now terrible. We probably also need to regenerate the floyd parser
(//floyd/parser.py) in order to bring it up to sync with the latest
compiler output.