Get printer test coverage to 100%.
This fixes one issue with the analyzer, several in printer,
and cleans up some dead code. It also makes sure that json5.g
pretty-prints to itself. We can't do this with floyd.g yet because
the lines would be much longer than 80 cols. Also leaves a TODO
for an improvement to the pretty printer for choices with actions.