Refactor grammar_test into mixin + base classes.
This moves the tests in grammar_test.py into a mixin class,
and then defines two concrete classes, Interpreter and Compiler,
that mix it in along with the one function that needs to be
implemented differently between them.

With this, code coverage of compiler.py gets almost back to 100%,
but there is one regression (difference?) where empty nodes aren't
being handled the same.

Also, five tests that pass with the interpreter fail with the compiler,
so I still need to fix those.