More work on handling grammars that are both left- and right-recursive.
After having investigated the mystery I was seeing in
`test_recursion_left_opt()` in grammar_test.py, I've concluded
that the code was mostly correct, except that there were some unneeded
lines and the mechanism for forcing left-associativity wasn't working
correctly. I've updated the code and TODOs to reflect that. Ideally
I'll figure out a way to make the behavior left-associative, perhaps
optionally, because that's probably more likely to be the desired
behavior.