use Optional[X] instead of X | None in type hints.
I was declaring types that were nullable as `X | None`, but
apparently that syntax was only added in 3.8. In looking into
how to fix this I first tried `Union[X, None]`, which works fine,
but then learned that `Optional[X]` is probably the more idiomatic
way to declare things, so I used that.