Add support for unicode character classes.
This adds a new ability to recognize unicode character categories.
If you add `\p{X}` as an expression to your grammar, it will match
if the next character's unicode character category is X. This
is a much cleaner alternative to using any and semantic predicates
that called _is_unicat. Once I can regen pyjson5 to use this version
of the grammar, I will remove _is_unicat as a builtin function.

This change also regenerates floyd/parser.py and updates a couple
things in todos.md to make it up-to-date.

This change is totally functional, but I still need to update tests and
address some coverage gaps.