Finish work on adding unicode character classes.
This tweaks the compiler to omit _leftrec if a grammar didn't need
it, and tweaks the generated code to add pragmas for a lack of coverage
if need be. At some point I should figure out if I can omit the
code that isn't covered if it's not possible to execute it.