Catch more errors via static analysis.
This change reworks the analyzer to be able to do actual static
analysis checks. Now any uses of unknown rules, variables, and
functions will be caught before the grammar is interpreted or
a parser is generated, and fewer problems will need to be caught
at runtime.

This (re-)adds an Analyzer class to do the analysis. The analyzer
can catch multiple errors at a time, and a single AnalysisError
exception is raised at the end of the static analysis pass containing
a list of the errors.

Unfortunately, the current parser implementation does not preserve
line and column number information after the initial parse, and so
we can't point at the locations in the grammar that actually have the
errors. This will have to wait until I can rework the parser to return
a structured node tree rather than just a JSON value.