Add support for positional variables in actions and preds.
With this change, you can now reference the value of an expr in
a sequence by its position in the sequence. For example, in the grammar

  grammar = "foo" "bar" -> $2

$2 refers to the "bar" literal. This is implemented by scanning any
preds and actions in a sequence for references to these variables and
then rewriting the AST to act as if the label had been explicitly set.

This change changes the grammar so that identifiers are allowed to start
with a '$', but those names are reserved for internal use and cannot be
explicitly defined.

For now, we continue to support named exprs (like "bar":v) for
backwards-compatibility and flexibility, but I suspect that I will
eventually remove that support so that there is only one way of doing
things.