Initial work on automatic filler insertion.
This is the first step to getting automatic "filler" insertion
to work ("filler" is comments and whitespace). This introduces
the concept of "pragmas" to the grammar, where a pragma is like
a rule but the name starts with a '%' and there is a limited set
of acceptable names.

There are four kinds of pragmas for now.
The first two are '%token' and '%tokens' and they declare which
rules will not have any filler inserted in them.

The second two are '%whitespace' and '%comment', and they define
what will be considered filler: basically filler = (whitespace | comment)*
and will be inserted before any literals or tokens or the end.

This CL only wires up the scaffolding for parsing and printing pragmas
and checking that tokens refer to valid rules. Nothing else is done
with the pragmas yet.

This CL is missing a few tests to bring the coverage up to 100%.