Get filler insertion working properly.
This implements the automatic insertion of whitespace and
comments as appropriate.

This also adds a grammars/json5_2.g which is a cleaned
version of json5.g that uses comments and a whitespace rule.

This made me realize that specifying comments or whitespace
rules but no tokens was okay; a grammar could just use literals
and ranges and be perfectly usable (see
tests.grammar_test.Interpreter.test_c_comment_style_with_range).