Rework how the style pragmas are specified.
This changes the grammar so that the comment_style and whitespace_style
pragmas are just specified to take identifiers instead of a hard-coded
list of strings, and then the analyzer checks to make sure the identifiers
are recognized. This will avoid having to change the grammar every
time a new style is added.