Rewrite filler insertion to be a bit more natural.
Prior to this change, we were doing filler insertion by rewriting
every node to fill as a parenthesized sequence. This would lead to
sequences with parenthesized sequences as nodes, instead of just
a single sequence with filler nodes inserted as need.

Now we'll insert filler into a sequence as needed, creating a flatter
AST.

In addition, this defines a '_filler' rule so that inserting filler
is just `['apply', '_filler', []]`.