Add primitive support for skipping different kinds of tests.
This adds decorators to grammar_test.py and printer_test.py that
can skip different kinds of tests, and defines two kinds of tests
for now: 'integration' for the slow tests that test the floyd and json5
grammars, and 'operators' for the operator tests (which are slow
unit tests). You specify what to skip in a comma-separated list of
strings in the environment variable SKIP.

This adds a --omit flag to `run coverage` that gets passed to
`coverage report` to not count or print coverage reports for
files that match given patterns.

And, this adds a -u/--unit flag to `run tests` and `run coverage`
that will set SKIP automatically to skip the integration tests (but
will still run the operator tests). This also automatically adds
the appropriate --omit flag to `run coverage` to skip reporting the
test files (since we're skipping tests, the coverage of them is
not really interesting any longer).