Rework line breaking in python_compiler.
This uses a new algorithm that just depends on nesting arrays
to indicate where things might be line-breakable, as opposed to
a bunch of enums giving directions.

This works well although it will not insert a final trailing comma
when breaking a list of args over multiple lines. This keeps the
formatting from matching ruff and black.