Fix a nasty bug in inlining parenthesized subnodes.
If you had a node that was a choice of two parenthesized subnodes,
the compiler would generate the same function name for both methods,
and the second one would overwrite the first one, causing parses to
fail. This CL fixes this by making interior subrules get unique
names even when the outer subrule is inlined and anonymous. The
downside is that the names become even more detached from the structure
of the node.