Fix some cases where there were still one-member sequences in the grammar.
The analyzer has a pass to rewrite a sequence with one element as just
the element. But, there were cases where we were missing those
optimizations, most notably when we would rewrite the AST to include
filler *after* doing the optimization pass. By fixing these cases,
I was able to simplify the inlining logic a bit.