v0.11.0 - rework API significantly
Rework the API significantly. Now the generated parser API is a single
`parse()` function and a data type for the return value (`Result`),
and the public API functions are called `compile()` instead of
`compile_parser()` and `generate()` instead of `generate_parser()`.

Add typing hints to everything.

Add lots more documentation of the API.