Invert failure checking in parser.
This changes the `failed` field in the parser to `ok` and inverts
its usage. I.e., where before we would've had:

    self._ch('a')
    if self.failed:
         return
    self._ch('b')

we'll now have

    self._ch('a')
    if self.ok:
        self._ch('b')

Checks that pass are ~10x as common as checks that fail, and so
this removes a lot more 'not' tests than it adds. This (somewhat
artificially) improves the coverage scores because now a successful
seq will execute every line in the method, rather than having N
different uncovered lines.

In theory, the branch coverage should stay about the same, though.

It's not clear if there's much of a performance benefit here, maybe ~1%.

Because the branch coverage is so awful after unrolling everything,
this also modifies the `run` script to make collecting branch
coverage a parameter we can turn off. Hopefully this will only be
temporarily useful until I can write a bunch more tests.