Tweak operator syntax, get statement coverage to 100%.
Previous operators could be any non-blank thing at all, which was
overly generous. This changes the syntax so that operators cannot
contain characters that are a legal part of an identifier (so,
not [_a-zA-Z0-9]. However, we need identifier-like things to be
legal when declaring the associativity of a recursive expression,
and so this tweaks the `%assoc` rule to accept either the new operator
or something of the form `ident '#' digits`, which is how the arms
of a recursive expression can be identified.