Rework method generation in python_compiler.py.
This gets rid of the type-based subrule names and switches to just
generating names of the form _{rule}_{i} where i ends up being a
consecutive series of integers for each rule. As a part of this
work this changes how the text of methods are passed around and
cleans up the approaches a fair amount.

This also changes the names of generated rules to start with '_r_' and
subrules to start with '_s_'; this means that we have separate
namespaces and prevent any collisions between rule names in the grammar
and generated subrule names.