Format to a line length of 67 instead of 71 in the formatter.
Expressions may now be indented 3 tabs deep instead of two, so we
need to wrap expressions when they are 67 chars long instead of 71.
In theory this might result in some lines that are too wrapped
(e.g., for an expression on a method line that is only indented two tabs)
but this doesn't seem to be a problem at the moment, so this is good
enough.