Move builtin functions to the top level.
This changes the implementation of the builtin functions
like `cat`, `join` so that they are top-level python
functions instead of methods. None of them required
access to the parser object state, and it's better if
they can't get it.

Also this fixes a bug in `test_saw` where I hadn't adjusted for
the 67 char change.