Resolve mypy issue w/ api.generate().
With the new JS backend, in `api.generate()` we can get a
generator either of the PythonGenerator class or the
JavaScriptGenerator class. On the JS branch, mypy seemed to
get confused about this, but after reformatting and merging
the issue appears to have gone away, so this has a proper
type declaration for it. TIL that in Python you can now
declare the type of a variable without also having to initialize it.