Update Floyd grammar and clean up docs.
This adds a new //grammars/floyd3.g file that should match
the language described in //docs/grammar.md. However, I have
not yet tested the grammar and there are surely mistakes.
The grammar also has a new AST representation and the interpreter
and compiler have not been updated for that yet, so you can't
do anything with this grammar even if there aren't any mistakes.