Get rid of comment_style, whitespace_style, get tests passing.
The tests were not passing when we tried to compile floyd2.g with
floyd2.g. It turns out that the regexp rules were wrong and we
weren't consuming input on whitespace.

Since I wanted to get rid of them anyway, this change gets rid
of comment_style and whitespace_style, and fixes the %comment and
%whitespace declarations for floyd2.g, which is now used for parser.py