Implement memoization for JavaScript, fix for leftrec & operator nodes.
The simple memoization algorithm doesn't work right when you have
left-recursive and operator nodes, because of the need to grow a
seed and update your result for a position several times until
your recursion is done. This change tweaks the algorithm so that
left-recursive and operator nodes are never memoized. There's probably
a way to be able to memoize things safely, but I haven't dug into
that yet.

This also rearranges the memoization cache so that instead of it
being a single dict/map with (rule_name,pos) as keys, it makes it
a two-level dict/map with Map(pos -> Map(rule_name, X)) instead.
This'll make it easy to discard unneeded cache entries when we get
around to doing that.