First pass at supporting global vars.
Everything works in the interpreter and the python generator, but
not the javascript generator yet. This is roughly feature parity
with the recent changes in glop.

This adds a new pragma, '%globals' so that a grammar can explicitly
declare the globals that it should be looking for. We should add
a check in parse that every expected global is being passed in.

Also, we still need to add new tests to cover this functionality.

Updates the JSON5 grammar to support the `_strict` option as requested
over in pyjson5-land.