Rewrite grammar_test.PythonGenerator to actually run the file.
This reworks the logic for the PythonGenerator test class to
actually spawn a subprocess to run the generated parser, rather than
using `exec()`. This is probably slower, but more realistic and brings
this closer to what we do for the JavaScriptGenerator. By refactoring
the logic to a common class there is a minimal difference between the two
Generator classes.