Rework pretty-printing to simplify, bypass analyzer.
Now we will only check grammars for syntax errors when pretty-printing.
This also simplifies the logic some, and changes things so that choices
are always split across lines rather than sometimes combined onto one line.
Might want to change this back in the future, but for now simpler is better.