Change how %whitespace, %comment, and %filler pragmas are handled.
Previously, if you wanted to refer to a %whitespace rule directly,
you'd have to refer to the underscore-prefixed version (_whitespace) instead.
This seems unnecessarily confusing; there's no reason to expose the internal
rules if we don't need to. So, now, you refer to %whitespace instead.

This also adds tests to ensure that you neither try to define a rule
starting with an '_' or refer to one.