Add `can_fail` property to Node.
Whether a node can fail or not is a property of the node plus the
grammar it applies to (b/c of the dependency on other rules in the
grammar).

When generating code, it's possible to optimize some things if you
know a node can't fail. Previously, in order to determine this you
had to query the node + the grammar. Now we set the property on the
node directly, so the AST can be more self-contained.