Derive the vars needed for decls ahead of time in js generator.
This adds a `.vars` attribute to the Seq AST node for variables that
are declared as local values in a Seq, and a `.local_vars` attribute
to the Rule AST that is used to keep track of the variables actually
needed to implement each rule in the generated language.

Ideally we would be able to declare each variable at the point of its
first use, but I don't currently see a way to do that without making
the text generation too context-sensitive (i.e., there's no way to
do that while implementing text generation using something like
StringTemplate that doesn't allow you to change state while generating
text), so for now we just collect all the variables that will be needed
ahead of time and declare them at the beginning of each function.

Even this requires more logic than is currently possible to do using
StringTemplate (I think), but it can't be done in the Analyzer because
the Analyzer doesn't know anything about the code generator, so I think
at best this needs to be done when the generator is initialized (and
before we start actually generating text).