Completely rework formatter.
This CL completely changes the formatter algorithm.

Previously I had used a weird recursive algorithm where we
gradually increased the "depth" to recurse through the tree;
the depth controlled whether to try and fit everything on one
line or not. It was really clunky and probably fairly inefficient.

Now we simply track how much space we have left on the line, and
more explicitly track when there are more than one ways to format
an object. When there is more than one way, we try formatting
each way with an "infinite" line length (so that child objects don't
try to use multiple methods). If none of those work, we then try using
the last (most verbose) method with the normal line length (so that
we can then try more verbose methods for the child nodes as well).