First work towards code generation via templates.
We need to move the code generation from code to templates, so that
the templates can be reused across different implementations of Floyd
in different languages. This means we need to pick a templating language.

After starting at things quite a bit, of the mainstream templating
systems StringTemplate (perhaps unsurprisingly) is the closest to
what we probably need. However, I don't want to wed myself to the
exact syntax and semantics of StringTemplate at this point; better to
roll my own as I figure out what I need.

For now I'm going to use something that is a mismash of datafiles,
StringTemplate rules, and the at-expr syntax from the Racket
programming language.

Instead of using StringTemlateGroups, we'll use a datafile where
each rule is either a string (potentially containing embedded at-exprs)
or a lisp-like expression. We'll keep the at-expr implementation
as minimal as possible and implement only the functions we need as
we go along to get the code generation to work.