Clean up Generator class structure, template formatting.
This splits the Generator base class in two, into one with the minimal
interface needed and one that can be used as a base class for the
hard-coded python and javascript generators.

This also cleans up how templates are parsed to reduce the number of
extraneous newlines being inserted. A newline at the beginning of a
string will be trimmed, and if an @expr evaluates to a string that
ends in a newline and is immediately followed by a newline in the
template, only one newline will be inserted in the result.