Make function signatures simpler.
This changes the expected signature for a function callback from
a variable list of positional args and one keyword arg (env) to
two fixed args: a list of parameters and an env. This reduces
the built-in signature checking you'd get for free from Python (
and means you can't really just bind an arbitrary Python function to a
name), but it also eliminates the reliance on that magic.