Rework how functions and fexprs are defined; cleanup.
By making functions and fexprs regular python objects, we can
now set more attributes on them. This change adds an 'is_fexpr'
attribute, which means that we can now define functions and fexprs
the same way and don't need them to be treated as two different
kinds of things in the eval loop.