rework var/identifier handling.
The 'e_var' AST node type was something of a misnomer, since it
was also used to refer to functions (and externs) as well as
local and outer variables. This change renames 'e_var' to 'e_ident'
and adds logic to track what kind of identifier it is, simplifying
code gen and interpretation.