Merge in the `at_expr_templates` branch.
This implements a declarative, template-based way of doing code
generation. The templates are roughly based on Racket's "at-exp"
syntax (used by Scribble, it's documentation tool), and use a
small Lisp interpreter as the underlying engine.

This mostly subscribes to ANTLR's StringTemplate philosophy of
using a functional language to generate text, but does not follow
ANTLR's belief that the template generation language should be
strictly limited in power. In particular, it's not clear if
StringTemplate can handle context-sensitive code generation, and
there's at least one situation (where you need to track whether
a variable has been declared already or not in a current scope)
that I can't think of how to implement in a non-context-sensitive
way, and that you can't push this logic into the model because
the model doesn't know about which variables might be being used
in the code generator.

The branch also contains the first full implementation of the
"datafile" language, a strict superset of JSON that uses as
little syntax as possible. The current way to specify templates
uses datafiles where templates are a dict of either multiline
strings (parsed as at-exps) or lisp functions (parsed w/ nested
arrays).

At this point there is a template-based Python code generator that
can exactly match the hard-coded code generator, but the JavaScript
implementation is still hard-coded. Next steps are to refactor the
template support to support two template-based code generators that
can share common logic.