Add ability to reuse other templates through inheritance.
This CL adds the ability for template files to reuse things from
other template files via inheritance. It changes the template file
format from a simple list of templates to a more complicated structure
with a top-level set of fields, one of which is 'templates'. Another is
'inherit', which takes a list of templates to inherit. Inheritance is
accomplished by merging template files, with the definitions from the
base files being overridden by the definitions in the derived template.
More than one base template is supported, but not currently used.

This change splits the Python template into a base template and a
Python-specific template. In theory the JS template should be able
to reuse everything from the base template.