Rename operators in hard-coded generators and switch python generator.
This switches src/pyfloyd/grammar_parser.py from having been generated
by the hard-coded python generator to having been generated by the
python templates.

This also renames the operators to 'o_' in the hard-coded generators
to match what the templates spit out. This makes it easier to
compare the output of the two; the only difference should be the
command line used to generate the output.