Implement support for variable length args in lisp user functions.
This adds the ability to declare a function that takes an arbitrary
number of arguments. To declare a function with two args, x and y, you'd write:

    [fn [x y] [strcat x y]]

Now, to declare a function that takes an arbitrary number of arguments, write:

    [fn args [length args]]

This also implements a `length` function to return the number of elements
in a list and (finally) adds a `cons` function as well.