Rework at-exp/datafile/formatting boundaries and collapse nodes.
This work several things:

1) It splits out the logic handling the at-exp language from the
   datafile generator into its own package so that you can use
   (and more importantly, test) the at-exp logic independently of
   datafiles and tests. The formatter functions also moved from
   the datafile_generator; even though they are technically kinda
   independent, in practice I'm always using the two linked together
   because we want the at-exps to be turned into format objects,
   not just a list of strings.

2) It changes how handling strings in templates works, to convert
   leading whitespace into a series of Indent objects. I.e., instead
   leaving '    foo' as a string, it gets turned into [ind 'foo'].
   In particular, this works across multiline strings.

3) This implements a bunch of FormatObj simplification and collapsing
   rules, as follows:
   1) the leading whitespace change in 2
   2) VLists, HLists, and Indents can take lists as args in addition
      to just strings and FormatObjs; this eliminates the need for the
      `_l` variants of the format functions.
   3) Nested VList, HList, and Indent objects are collapsed where possible,
      e.g., [vl [vl foo]] becomes [vl foo].
   4) Adjoining Indents are merged together, e.g., [ind foo] [ind bar] becomes
      [ind foo bar]. Note that this happens across empty lines as well, so
      [ind foo] '' [ind bar] becomes [ind foo '' bar].
   5) HLists that are just a series of strings get strcat'ed together into
      a single string, e.g. [hl 'foo' 'bar'] becomes just 'foobar'.

This should get us to a formatter tree of objects that is close to ideal.
In most cases these optimizations aren't really necessary, as the code
generation ends up being the same and this is just slowing things down
for aesthetic purposes, so at some point it might make sense to make this
happen only when the the tree is the desired output, rather than formatted
text.

A more significant improvement is that this theoretically makes the
output independent of the indentation used in the templates, as all
of the hard-coded indentation whitespace in strings has been converted
into Indent objects. I tested this a bit by hand and it seemed to work
correctly, but there is no programmatic way to tell the code generator to
use different amounts of indentation in the templates and output yet,
and things are probably pretty buggy if the indentation isn't consistent
in the templates.

And everything still needs a lot more tests.