Rework command line args, part 1.
Code generation was already using templates by default, but this
changes the command line args to reflect that. For now, this gets
rid of the somewhat ambiguous `-l`/`--language` flag in favor
of a `-g`/`--generator` flag. It also redefines the `--js`/`--javascript`
and `--py`/`--python` to specify datafile templates instead of
a particular generator.

This also changes the flags for outputting a format tree of objects;
the flag is now `--ft`/`--format-tree`. You can also use the `--as-json`
flag in conjunction with `--ft` to get the tree as a JSON object instead
of a lisplist-formatted datafile.