Rework command line args, part II.
This adds the ability for a grammar to specify a preference for
memoization using `%extern memoize -> true` (even if the value
is not otherwise used in the grammar itself). The datafile generator
will now try to use the value from the grammar if it has not been
given an explicit memoization option.

This also renames the `-D`/`--define` flags to `-E`/`--externs`, to
make it a little clearer which things the flags will affect.

This also removes the `--memoize`/`--no-memoize` flags in favor of
now specifying things via the `--externs` or `--generator-options`
flags.

This change also tweaks `run` to rely on the memoization defaults
in the grammars.

It looks like a bug has crept into the formatting code, as
the datafile parser no longer matches the `ruff`-formatted version
of it. I'll deal with that in a separate change.