Fix formatting of Saws.
The existing algorithm for formatting Saws didn't work right. The
problem was that when splitting across multiple lines, you need two
different new lengths to use, one for the first part of a triangle
and one for the second part. For example, if you had

    foobar(1, 2, 3)[long_var]

This would get parsed to
    Saw('foobar',
        Triangle('(', Comma('1', '2', '3') ')'),
        Triangle('[', 'long_var', ']'))

Across multiple lines, this should get formatted to:

    foobar(
       1, 2, 3
    )[long_var]

and the calculation for whether the first  Comma expression fits needs two
different lengths, one to account for "foobar(" and one for "    ".
This couldn't be done with the existing implementation, but worked
before because we would modify the triangles to account for the
two differences.

The new implementation takes an approach similar to the old one,
but without modifying the existing objects passed to Saw; instead the
Saw decomposes the triangles into individual objects and merges them.
So, instead of the above leading self.objs to be

   ['foobar', Triangle('(', Comma(...), ')', Triangle('[' ... ']'))]

it becomes:

   ['foobar(', Comma(...), ')[', 'long_var', ']']