Generate built-in functions from a datafile.
Prior to this change, the "built-in" functions were being maintained by hand
in *eight* different places:
- docs/functions.md (the documentation for each function)
- src/pyfloyd/grammar.py (the list of function names)
- src/pyfloyd/interpreter.py (the grammar-interpreter's implementation)
- src/pyfloyd/lisp_interpreter.py (the lisp-interpreter's implementation,
      used by the datafile generator)
- src/pyfloyd/python.dft
- src/pyfloyd/javascript.dft
- src/pyfloyd/python_generator.py
- src/pyfloyd/javascript_generator.py

Ideally you'd only need one implementation for each target language,
and you'd be able to re-use the target implementations in the host floyd
implementation. You'd also be able to generate the documentation from
the implementation.

This change implements the machinery to do that. Now the functions are
defined in a single `functions.df` datafile that contains the function APIs
and the Python implementations. I've written additional templates that will
generate a standalone `functions.py` file and a `functions.md` file, and
updated the base.dft and python.dft templates as well as the
datafile_generator to be able to generate the functions from the datafile.
Among other things, this adds the ability for templates to specify datafiles
that should be read to load data that can be used by the template.

This change does not yet implement the JS definitions in a separate datafile.
I'm undecided as to whether to keep those definitions alongside the
python ones in the datafile, or to keep them in the JS code generation
template (or perhaps even in a separate file somewhere else).

This change also does not modify the hard-coded generators to use the
functions; I'm not sure that it's worth it given that they are on their
last legs.

One of the built-in functions is `dedent`, but up till now it had not
been implemented properly. This change moves the datafile/api.py
implementation into the template and changes the datafile/api.py
code to import functions.py; this makes the datafile module no
longer fully independent of the rest of pyfloyd, but that's probably
okay. Eventually we'd like to get to the point where the datafile/api
code is generated from templates as well, but we're probably a good
way off from being able to do that.