Rework how generator_options are processed.
This changes things so that values in `generator_options` are
never modified, and values in generator.data are rarely modified
(only to provide default values for indent and line_length if they
aren't specified at all).

This also gets rid of the argparsing argument options in datafile;
I forgot I was using those and ended up being really confused for a
while, so I've decided they are more trouble than they're worth.