Add Hang(), Wrap() format objs and fix command line wrapping.
When printing the command line used to generate a file from
@generator_options.command_line, there was no way to split the
command line across multiple lines. This change adds two new
FormatObjs to make that possible.

`Hang()` will take a list of objects and format them across multiple
lines, with the second and subsequent lines indented to line up with
the second object.

`Wrap()` will take an object and format it across multiple lines,
wrapping it with text on both sides.