Get Values tests passing for Go code gen.
Among other things, this fixes an issue in the analyzer where
we would indicate that a variable that was being used as a label
in an outer scope was being marked as if it still needed to be
declared as a local variable.

This also adds panic/recover handling in parse_with_exception to
be able to handle bad predicate values.