Rework how quals work.
In the floyd grammar, a "qual" is a "primary" expression followed by one
or more postfix expressions, where each postfix expression is either
a function call (e_call) or a object dereference (e_getitem).
This showed up in the AST as a node with a list of children,
the first element being the primary expression and the remaining elements
the postfix expressions.

This had two problems. First, it was awkward to figure out how to
interpret or execute the node, because you'd have to look at multiple
children to try and figure out what the right thing to do was, e.g.,
is the primary expression an identifier and the first postfix expression
a call? Then it's a function call.

Second, in some cases you can't actually generate the right code
purely from looking at one of the children. For example, in the
current set of Go templates, because we don't have solid typechecking,
we can't correctly distinguish between a primary expression that is a
list literal and a variable containing a list, and Go won't let you type
cast the former directly. So, you have to instead use a wrapper function
to handle the potential conversion.

Perhaps more obviously, in a language that doesn't actually have
postfix expressions (like Lisp), the structure was going to be even
harder to handle.

This change changes things by rewriting the AST so that you have a
tree of "call_infix" and "getitem_infix" nodes, each of which takes
two arguments. Everything becomes much simpler and cleaner.

With this change, the last Go tests pass (apart from bignum support).