More checkpointing.
Fix `outer` variables to return the correct type.
Fix at_exp and datafile grammars to use `list()` instead of `[]`.
Change `run regen` to use `flc` and `fld`, and to print out the
regen command if it fails.
Modify the way errors are propagated so that you can dump the
AST even if there were semantic errors found; this gets rid
of the AnalysisError exception, and now the analyzer will always
return the grammar and not "fail".
Now regen passes and all the tests pass.