Merge simple cases of list and tuple types.
Now if typechecking gets `list[any]` and `tuple[]`, it'll infer
the type `list[any]`. This allows us to use `[]` as a literal
in the grammar and still the type unified when we use that and
a function like `cons()` that will always return a list type.

There's probably more work to do here to better unify different
kinds of tuples and lists, but this'll do for the moment. We
definitely need tests, too.