Add type information to node vars and local_vars.
This changes the AST and the generators so that `Seq.vars` and
`Rule.local_vars` are now mappings from var name to type descriptors.

This allows a generator to properly declare the type of a local variable.

Only the Go code generator uses this currently, and unfortunately it
only uses the information for Rule.local_vars, and not Seq.vars. This
is because `Rule.local_vars` is used for things like local list
indices, where everything is narrowly typed. For `Seq.vars`, we're
typically assinging to and from `Parser.val`, which is typed as `any`.

And, Go's typing prohibit casting directly back and forth from
`[]T` to `[]any`; you have to manually construct the lists and
copy things over each time. This is expensive and makes the code
kinda ugly, so the code gen just sticks with `any` for all node types.
This is unfortunate, but it seems like the correct tradeoff at the moment.
I need more experience with other typed languages to see if Java
et al are going to have similar problems.