Bump version to v0.25.0.dev0, update TODOs.
At this point I consider the Go code gen to basically be done,
and the current internal support for typechecking to basically be done,
until I have evidence that there's more to be done that could improve
things.

We will now catch at parse time cases where the built-in functions are
called with incorrect types, and catch when something is passed to
the pred (`?{}`) operator that isn't a boolean.

There is still the possibility for runtime errors for things like
calling `atof()` with a string that isn't a number; we haven't defined
how to handle that yet.