Go back to using TypeDescs directly.
Initially I was using a custom class to destructure type strings
and storing those in the AST. However, that meant you couldn't
convert the AST directly to JSON (without providing a custom
decoder, which I didn't want to do). So, I started storing type
information in regular dicts. However, I wasn't happy with the loss
of type checking.

I realized I could change the custom class to be a subclass of dict,
and then it would work correctly w/ the json module while still
providing strong typechecking. This change implements a `FixedDict`
base class to do that work (and moves attr_dict.py to custom_dicts.py
to contain both AttrDict and FixedDict), and then changes TypeDesc
to subclass that.