Make it possible to pass external functions to a parser.
We need to be able to customize the behavior of the parser beyond
what is supported by the floyd language. To this end, you can now
declare an `%extern = foo -> func` in your grammar that says that
there will a function passed in that you can invoke via the name `foo`.
This function can shadow an existing function, so that you can
override its behavior.

This change implements this for the Floyd grammar itself; there is
now a built-in `node()` function that takes a list of args and
returns it as a list, but it can be overridden to generate custom
syntax nodes.

The external function will be passed the parser object as its first
argument, so it can query the parser's state.