Regen grammars, use `node()` for floyd.g.
It looks like I lost something in the Node() cleanup and something
is no longer being inlined (possibly a `not` node?). All the tests
pass, so I'll look into this down the road.

This switches all the Floyd AST nodes to be created by the `node()`
function; this will allow me to add more state to the nodes without
needing to modify the grammar (as much).